requirements.txt
Adafruit-Blinka==8.32.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-colorsys==2.0.17
adafruit-circuitpython-requests==2.0.5
adafruit-circuitpython-typing==1.10.2
Adafruit-PlatformDetect==3.60.0
Adafruit-PureIO==1.1.11
asgiref==3.7.2
Django==5.0.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
pillow==10.2.0
pyftdi==0.55.0
pyserial==3.5
PythonTurtle==0.3.2
pytz==2024.1
pyusb==1.2.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
